fix: Wait for L1 to L2 msg sync before claiming (#12386)
In order to claim an L1 to L2 msg on L2, we need to wait 2 L2 blocks to
be mined _after the archiver has synced the msg_. We were waiting 2 L2
blocks since the message was sent L1, so if this happened right before
the end of an L2 slot, the archiver would not get to sync it in time, so
waiting for 2 blocks meant we only waited for **one** block after it got
synced.

This adds a wait for the msg to be synced in the node before waiting for
the two L2 blocks. The code that waited for the L2 blocks was repeated
in three different places, and I didnt attempt to refactor it here, so
the fix is also repeated three times.

Fixes #12366