feat: nullify just-added notes (#12552)
Back when PXE was a service processing all blocks, we'd remove nullified
notes as we saw their nullifiers. This later got changed, and as of
https://github.com/AztecProtocol/aztec-packages/pull/10722 we remove all
nullified notes whenever we 'sync' notes. However, note syncing is
becoming less and less a part of PXE, and as of
https://github.com/AztecProtocol/aztec-packages/pull/12391 we even
deliver notes _outside_ of the PXE-led note syncing process (whenever we
complete partial note). This causes problems because we end up adding
notes, failing to realize they've been nullified and then returning them
via `get_notes` (which is what causes some tests in
https://github.com/AztecProtocol/aztec-packages/pull/12391 to fail). The
next time a contract function is run we'll do note syncing again and
they'll be then removed, but we did have a full fn call in which they
were available.

This PR makes it so we always check if newly-added notes have been
nullified, and remove them if so. I also added some explanations re. why
we're doing things this way, created some follow-up issues (mostly
#12550 and
https://github.com/AztecProtocol/aztec-packages/issues/12553), and
inlined `produceNoteDaos` to have the whole thing happen in a single
place. I think it's now more readable but potentially slightly large -
perhaps this will improve as we split `PxeOracleInterface` in multiple
files or modules.