chore: avoid syncing recursive proofs between repositories (#12769)
All of the recursive honk tests are not really useful in the noir repo.
We never actually generate any proofs for the contents of the
`test_programs` directory so having changes to the proof format going
through the sync mechanism just adds churn for no reason.

I've then added a folder to `barretenberg/acir_tests` where the crypto
team can add any test programs which shouldn't be synced across to the
noir repository. Anything where the inputs are generated from bb should
live within this folder which is merged with the `acir_tests` directory
during bootstrapping.