feat: A strict scheduler for tests/benchmarks. Currently not enabled. (#13561)
This investigates the use of an alternative to `parallel` in the
`parallelise_strict` script. Every test is exclusively allocated and
pinned to the requested number of cpus (default 2).
Sadly at present, it's demonstrated that even with exclusive cpu access,
we get **different performance** profiles when other cpus are busy
(accounting for logical cpus already).
This means at present at least, we can't "mix in" the benchmarks with
the rest of the tests.
Needs further investigation, as there is no technical reason the
performance should not be stable in this model.
For now I've left the existing parallel implementation in place, strict
scheduling can be enabled with `STRICT_SCHEDULING=1`.
There were a few other improvements that made their way in:
* Better signal handling and cleanup, particular around docker
containers.
* Unifying around `docker_isolate` script and making it a "first class"
test framework request. Cleans up docker gubbins from the `run_test.sh`
scripts for:
  * acir browser tests
  * yarn-project tests
  * end-to-end tests
* Create a new `run_compose_test` to unify how we run compose tests.
Cleans up docker compose gubbins for:
  * boxes tests
  * end-to-end tests
* Passing of env vars to the testing framework with e.g.
`<hash>:TIMEOUT=15m:ISOLATE=1:CPUS=16`
* e2e tests were sometimes forgotten to be added to the test_cmds
function. As part of supporting the above, I've also updated it to be
more of an "exclusion" rather than "inclusion" model. Also much cleaner.
* exclude `bb/deps` from bb format check as cli11.hpp was absolutely
killing its performance.
* `cache_content_hash` tried to error out in CI on changed file
detection. But it saw things actually just stumble on as I don't think
exit code is properly handled in a `blah=$(cache_content_hash)` context.
Back to a warning.