feat!: AVM DebugLog opcode should be (mostly) a no-op unless doing verbose "client-initiated-simulation" (#14085)
Unless initiated from `node.simulatePublicCalls()` or from TXE, with
log-level verbose, DebugLog should be a no-op.

It is a no-op, but it resolves memory address operands to avoid the need
to special case things in C++. Special casing would be "this instruction
has memory offset operands but we only _sometimes_ want to resolve and
tag check them and generate events".

It does not do any slice memory stuff when in no-op mode.

Co-authored-by: AztecBot <tech@aztecprotocol.com>