test: merging e2e_outbox with l2_to_l1 test (#14246)
Made no sense to have these 2 tests separated as they test the same
thing (L2 to L1 messaging). Decided to tackle this tech debt task now as
it confused me when debugging a related issue (incorrectly assumed a
feature is untested while the test was hidden in the outbox test).