feat: `PublicImmutable` initialization optimization (#14608)
Upon `PublicImmutable` initialization we wrote a value to a designated
initialization slot in public storage that later on was used to
determine if the value is already initialized or not. In this PR I
replace the public write with emission of a nullifier which yields the
following savings:
1. Lowered DA cost by 1 field as emitting a nullifier consumes 1 field
but writing to public storage 2 (1 extra field for the tree index),
2. cheaper reads as it's cheaper to verify nullifier inclusion compared
with public tree value check as nullifier tree is much smaller (hence we
need less hashing to perform the merkle proof) - note that this applies
only to `read` method as `read_unsafe` method doesn't perform
initialization check.

---------

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>