chore: Remove `HasPlookup` from `bigfield` and `biggroup` classes (#14540)
Please read [contributing guidelines](CONTRIBUTING.md) and remove this
line.

Since `StandardCircuitBuilder` has been removed, we want to get rid of
the boolean `HasPlookup`. In this PR, we get rid of `HasPlookup`
conditional branches in the `bigfield` and `biggroup` class. Summary of
the PR:
- Removes `HasPlookup` keyword from the bigfield` and `biggroup` classes
- Separate test suite for the `biggroup` tests for the `secp256k1` curve
since we have custom functions for `secp256k1`. These tests were not
being triggered in the current setup (not even with
`UltraCircuitBuilder`.