fix: comptime friendly Serialize imp of BoundedVec (#14817)
Our implementation of `Serialize` for `BoundedVec` was not comptime
friendly:
<img width="1731" alt="image"
src="https://github.com/user-attachments/assets/4cf22749-2e35-4514-bfd4-b504b5d677fe"
/>

We need this to work if we want to pass BVec as arg to a contract
function (there Serialize is used for the args hasher stuff).

This implemenation doesn't feel right as it calls serialize on a
zeroed-out items. But I don't know what to do. Submitting this PR to get
some feedback.