fix: Rollup invalid archive error (#14785)
Fixes multiple `Rollup__InvalidArchive` errors that were popping up on
tests since https://github.com/AztecProtocol/aztec-packages/pull/13700.

Example from [this run](http://ci.aztec-labs.com/c82b760acf321da6):
```
09:42:54 [09:42:54.569] VERBOSE: sequencer Unable to build/enqueue block Rollup__InvalidArchive(0x2fdc940e38cff0a9fd112204b3ad9b7e75c3efaf9495dd7aeaecbcc8d2bba20f, 0x18674a63a7dd5a173cff9633512b7e5bd6b5536ea0b7c96929dcc68473c31e45)
```

The cause for this was the removal of the `canProposeAtNextEthBlock`
check
[here](https://github.com/AztecProtocol/aztec-packages/pull/13700/files#diff-1c2150b6e6b122e3a184b1132b8d6946e4e27fefe53381035810638fabcebe9cL384).
This call to the rollup contract was removed since we were now manually
checking the proposer using the epoch cache (which is a good thing!),
but we lost the check for the latest archive root.

This meant that, if the sequencer started building _before_ the archiver
was caught up to the previous block, it would build off an older chain
tip, and only get rejected during block publishing. Note that in an
actual setting, it should actually fail to collect attestations, since
the proposal should be deemed invalid by validators due to a wrong
parent, but most of our e2e tests don't have a committee.

This PR makes the sequencer check that the archiver has synched to the
previous L1 block before proceeding. Note that L1 missed slots would
cause issues here, but we can address that when we get to
#[14766](https://github.com/AztecProtocol/aztec-packages/issues/14766).
In addition, this PR restores the early call to
`canProposeAtNextEthBlock` to be on the safe side before we start
building.