refactor: ordering events by commitment index (#14919)
To be able to return private events in the order in which they were
emitted we stored `logIndexInTx`, `txIndeInBlock` and `blockNumber`
along with the event and then we ordered based on these 3 values. This
solution started to be problematic with offchain message delivery as
when an event is delivered via the offchain message channel it doesn't
have an index "compatible" with the private log index since it's a
completely separate information channel.

However with [this
PR](https://github.com/AztecProtocol/aztec-packages/pull/14634) we
commit to events and we insert the commitment to the nullifier tree.
This gives us a global ordering we can use.

In this PR I modify the code such that we order by this event commitment
index.