fix: Limit the number of function names stored in archiver (#14989)
Limits the number of function selector-name pairs stored in memory in
the archiver, and the maximum length for names. After the limit is hit,
no more values are stored, but the method does not fail. Also enforces
limits on the API method to prevent massive requests to the node.

Removes the unused `address` parameter in the `register` function.

Fixes #14257