fix: specify generic argument N of several structs (#14967)
For https://github.com/noir-lang/noir/pull/8686

When returning
`dep::aztec::context::call_interfaces::PrivateStaticCallInterface` (and
others) from some functions, the generic `N` argument wasn't specified.
This will soon cause a compile error (like in Rust).

I assumed the `N` argument is the size of the return type (I tried with
zero first and it said it didn't find a `Deserialize<0>` impl for some
types, so I figured the size in fields of the type was needed), but let
me know if that's not how it should be done.

Given that `size_in_fields` is now used in a couple of places, I decided
to move it to its own file/module.