feat: Slasher uses separate private key (#15712)
Adds a new config variable `SLASHER_PRIVATE_KEY` for the slasher client.
If not set, and L1 mnemonic is not set, then it derives it as the second
account of the mnemonic (first one is for the sequencer publisher).

Rationale for this change is having separate nonce management for
`Rollup.propose` blob txs, and non-blob txs
`SlasherFactory.createSlashPayload`, since blob and non-blob txs have
different replacement policies.

**Pending**: Update spartan scripts to generate and inject these
variables in all validators, otherwise slashing will be disabled.

---------

Co-authored-by: spypsy <spypsy@outlook.com>
Co-authored-by: Alex Gherghisan <alexghr@users.noreply.github.com>