fix: resolve DinD dns namespace resolution  (#15663)
When running docker in docker (DinD), the dns resolution file
(/etc/resolv.conf) is copied to the cluster and may be a private DNS
resolver that will not allow the container to reach outside.

- Added a cluster post-run patch to the bootstrap to use public DNS if a
env tag is enabled (KIND_FIX_DNS). This will restart the coreDNS
container with public DNS.
- This PR does not affect the underlying /etc/resolv.conf nor the
kind-config.yaml so handrolling the cluster will not be affected.
- Added documentation to enable in the quickstart.md.

---------

Co-authored-by: danielntmd <danielntmd@nethermind.io>
Co-authored-by: Alex Gherghisan <alexghr@users.noreply.github.com>