chore: toggle whether or not we use multithreaded or row-skipping compute_univariate (#16079)
In [14272](https://github.com/AztecProtocol/aztec-packages/pull/14272)
and [15841](https://github.com/AztecProtocol/aztec-packages/pull/15841),
a row-skipping functionality was incorporated in the process of
computing the round univariates in sumcheck. This passed over a
multithreaded implementation, which was tailor-made for the AVM, causing
a performance hit for the AVM.

We combine these two to maintain the AVM performance.

---------

Co-authored-by: notnotraju <raju@aztec-labs.com>