chore: audit ECCVM msm relation (#16532)
Audit of the MSM relation.

There were a couple of columns that I think were underconstrained. I have made the MSM relations more
robust/clear as to why they are correct.

Aside from a few changes to the relations, this PR mostly involves a lot of documentation. The most important part of this is re: the multiset equality check. The description of _why_ this suffices to correctly constrain the q_add columns is more transparent now.

---------

Co-authored-by: notnotraju <raju@aztec-labs.com>