fix: Check correct proposer in block proposal validator (#16825)
Block proposal validator would check that a proposal was for the current
or next slot, and the proposer was one of the current and next. But not
that if the proposal was for the current slot, the proposer was the one
for for the current lost as well.

Credit goes to @paperclip-minim for spotting it.