refactor!: fixing illogical code locations (#16891)
With [my recent
PR](https://github.com/AztecProtocol/aztec-packages/pull/16091),
location of the event and note emission related code became illogical.
In this PR i reshuffle it around.

I perceive it as quite important to have this merged relatively quickly
as that will allow devs to deal with 1 migration instead of 2.