fix(avm)!: unconstrained reads in debuglog (#17061)
Debuglog is not connected to the memory trace, so we can't generate events. The objective of this PR is to enable that while trying not to break the MemoryInterface.

I'm using type-checked casting `dynamic_cast` and adding an unconstrained read to the subclass `Memory`. Still evil, but a bit less evil than changing the interface itself.

In the circuit: I re-enabled the memory selectors check, and check circuit passes with `-d` (which should trigger debug logs).

Closes #16759.