chore!: ECDSA audit part 4 - tests (#17301)
### 🧾 Audit Context

This PR is the last in the ECDSA audit thread, it improves testing.

### 🛠️ Changes Made

Test refactoring and improvement.

Vk change comes from adding the constraint that the public key supplied
in an ECDSA verification is not the point at infinity.

### ✅ Checklist

- [ ] Audited all methods of the relevant module/class
- [ ] Audited the interface of the module/class with other (relevant)
components
- [ ] Documented existing functionality and any changes made (as per
Doxygen requirements)
- [ ] Resolved and/or closed all issues/TODOs pertaining to the audited
files
- [ ] Confirmed and documented any security or other issues found (if
applicable)
- [x] Verified that tests cover all critical paths (and added tests if
necessary)
- [ ] Updated audit tracking for the files audited (check the start of
each file you audited)

### 📌 Notes for Reviewers

N/A