refactor: rename ClientIVC to Chonk throughout codebase
The ClientIVC protocol needed a name. It builds on MegaHonk (UltraHonk with deferred non-native field ops) to allow folding circuits for more efficient multi-circuit proving. An obvious candidate is the somewhat boring ClientHonk. But through a refinement approach of @lucasxia01 saying Chonk and Ariel blessing it and using some throwaway AI art I made as propaganda, Chonk became a contender. If this PR gets merged, you know it won :)