refactor(avm)!: refactor PublicTxResult (#18124)
Refactors PublicTxResult in TS and C++, though not as much as I would've liked. This is only a step forward.

## NestedProcessReturnValues

Could not remove `NestedProcessReturnValues`: apparently there are parts of AztecJS that need to introspect these values so we will need to collect them.

I talked with @Thunkar and @nventuro  about this, and apparently they are probably they are only really needed in client initiated simulation mode. This is important because they might be too costly to gather for blockbuilding/etc. I also made it clear that these values CANT be trusted. They are in principle NOT constrained and will never be.

## Other

* Most information should still be collected in C++.
* Starts setup to return it from C++ to TS. MP and Zod setup seem to be broken so I have to do it little by little in a follow up PR.