feat: panicking on direct calls to external funcs (#18126)
Fixes F-27

Given `#[private] fn foo()` we want for `foo()` to issue a compilation
error. To achieve that we need to replace the function's body with a
`static_assert(false)`.

Given that the #[external(...)] macro is being processed before the
#[aztec] macro we can do the following:
1. In the `#[external(...)]` macro invocation I register the function
into 3 global variables based on the external type (`private`, `public`
or `utility`),
2. then in the `#[aztec]` macros for each of these functions I generate
a new function with `__aztecnr_internals__` prefix by taking the
original function's args and body, doing the same transformation on it
as we do now for external funcs and then I inject these new functions
into the contract,
3. then I replace the body of the original function's
`static_assert(false, "yo, go check the docs how to actually call a
function");`
4. for the functions that are currently injected by the `#[aztec]` macro
(e.g. `sync_private_state`) I would not apply the `#[external(...)]`
macro attribute as is currently done but instead I would just directly
call the "transformation function" on it that is also used in step 3 and
then inject the function.

Note that this PR is currently not mergeable as I needed to disable
caching in BB to get a fresh AVM transpiler + there is a broken test.
Both of these are (most likely) unrelated to this PR so I think this can
be already reviewed.