feat: `SinglePrivateImmutable` (#18792)
Partially addresses F-187 (to be fully tackled `SinglePrivateMutable`
needs to be implemented as well).

Just like `PrivateImmutable` but allows only for 1 instance per contract
instead of 1 instance per contract-user pair.

That 1 instance invariant is achieved by having a nullifier that has
only the storage slot in it's preimage. This currently leaks that
initialization happened - tackling this is a TODO.