feat: priority fee calculation strategies (#18910)
This PR adds an L1 priority fee analysis tool for nodes running in
fisherman mode.

Instead of clearing all queued requests from the publisher, fisherman
nodes will now calculate a priority fee given a specific strategy, then
wait for the ethereum block to be mined and make an analysis on how
competitive the fee we calculated was.

This can be done with multiple strategies.

Analysis is logged & pushed to our metrics on each strategy as attribute

Fixes
[A-307](https://linear.app/aztec-labs/issue/A-307/observability-and-metrics-on-l1-gas-fee-calculation-in-fisherman-mode)

Used Claude a lot for the analysis parts since they don't affect
production node code paths.
Not for the update on L1 tx utils that now uses a strategy object for
calculations (making it easy to swap. could possibly be made into a
config value)